--- src/chat/determineChatContainer.ts
@@ -1,14 +1,0 @@
-import { NamedNode } from "rdflib";
-
-export function determineChatContainer(
-  invitee: NamedNode,
-  podRoot: NamedNode
-): NamedNode {
-  // Create chat
-  // See https://gitter.im/solid/chat-app?at=5f3c800f855be416a23ae74a
-  const chatContainerStr = new URL(
-    `IndividualChats/${new URL(invitee.value).host}/`,
-    podRoot.value
-  ).toString();
-  return new NamedNode(chatContainerStr);
-}
