+++ src/index.ts
@@ -5,12 +5,6 @@ const authn = solidLogicSingleton.authn
 const authSession = solidLogicSingleton.authn.authSession
 const store = solidLogicSingleton.store
 
-const chat = solidLogicSingleton.chat
-const profile = solidLogicSingleton.profile
-const typeIndex = solidLogicSingleton.typeIndex
-const inbox = solidLogicSingleton.inbox
-const container = solidLogicSingleton.container
-
 export {
   findAclDocUrl,
   setACLUserPublic,
@@ -21,37 +15,52 @@ export {
 export {
   ensureTypeIndexes,
   loadTypeIndexes,
-  getRegistrations,
   registerInTypeIndex,
   loadIndex,
+  ensureOneTypeIndex,
+  putIndex,
+  makeIndexIfNecessary,
+  loadIndexes,
+  getTypeIndex,
+  getRegistrations,
   loadTypeIndexesFor,
   loadCommunityTypeIndexes,
   loadAllTypeIndexes,
   getScopedAppInstances,
-  getAppInstances
+  getAppInstances,
+  suggestPublicTypeIndex,
+  suggestPrivateTypeIndex,
+  registerInstanceInTypeIndex,
+  deleteTypeIndexRegistration,
+  getScopedAppsFromIndex
 } from './typeIndex/typeIndexLogic'
 
 export { offlineTestID, appContext } from './authn/authUtil'
-export { createInboxFor } from './inbox/inboxLogic'
+export { createInboxFor, getNewMessages, markAsRead } from './inbox/inboxLogic'
 export {
   followOrCreateLink,
   loadOrCreateIfNotExists,
   recursiveDelete,
-  setSinglePeerAccess
+  setSinglePeerAccess,
+  createEmptyRdfDoc
 } from './util/utilityLogic'
 
 export {
   ensureLoadedPreferences,
   loadPreferences,
-  loadPreferencesNEW,
+  loadPreferencesThrowErrors,
   loadProfile,
-  loadProfileNEW
+  loadMe,
+  getPodRoot,
+  getMainInbox,
+  findStorage
 } from './profile/profileLogic'
 
-export { SolidLogic } from './logic/SolidLogic'
 export { getSuggestedIssuers } from './issuer/issuerLogic'
+
+export { SolidLogic } from './logic/SolidLogic'
 export { AppDetails, SolidNamespace, AuthenticationContext } from './types'
 // solidLogicSingleton is exported entirely because it is used in solid-panes
 export { solidLogicSingleton } from './logic/solidLogicSingleton'
 export { UnauthorizedError, CrossOriginForbiddenError, SameOriginForbiddenError, NotFoundError, FetchError } from './logic/CustomError'
-export { authn, authSession, store, chat, profile, typeIndex, inbox, container }
+export { authn, authSession, store }
