+++ src/profile/ProfileLogic.ts
@@ -1,77 +1,16 @@
+import * as $rdf from 'rdflib';
 import { NamedNode } from "rdflib";
-import { solidLogicSingleton } from "../logic/solidLogicSingleton";
-import * as debug from "../util/debug"
 import solidNamespace from 'solid-namespace';
-import * as $rdf from 'rdflib'
-import { CrossOriginForbiddenError, NotFoundError, SameOriginForbiddenError, UnauthorizedError, FetchError } from "../logic/CustomError";
+import { CrossOriginForbiddenError, FetchError, NotEditableError, NotFoundError, SameOriginForbiddenError, UnauthorizedError, WebOperationError } from "../logic/CustomError";
+import { solidLogicSingleton } from "../logic/solidLogicSingleton";
 import { AuthenticationContext } from "../types";
+import * as debug from "../util/debug";
 import { followOrCreateLink } from "../util/utilityLogic";
+import { differentOrigin, suggestPreferencesFile } from "../util/utils";
 
+
+const store = solidLogicSingleton.store
 const ns = solidNamespace($rdf)
-
-export async function loadProfile(me: NamedNode): Promise<NamedNode> {
-    /*
-    // console.log('loadProfile cache ', solidLogicSingleton.cache)
-    if (solidLogicSingleton.cache.profileDocument[me.value]) {
-    return solidLogicSingleton.cache.profileDocument[me.value];
-    }    @@ just use the cache in the store
-    */
-    console.log('loadProfile  me ', me)
-    const profileDocument = me.doc()
-    try {
-        await solidLogicSingleton.store.fetcher.load(profileDocument);
-        return profileDocument;
-    } catch (err) {
-    const message = `Cannot load profile ${profileDocument} : ${err}`;
-    throw new Error(message);
-    }
-}
-
-export async function loadPreferences(me: NamedNode): Promise<NamedNode> {
-    
-    await loadProfile(me) // Load pointer to pref file
-    const preferencesFile = solidLogicSingleton.store.any(me, ns.space('preferencesFile'), null, me.doc());
-
-    function differentOrigin(): boolean {
-    if (!preferencesFile) {
-        return true;
-    }
-    return (
-        `${window.location.origin}/` !== new URL(preferencesFile.value).origin
-    );
-    }
-
-    if (!preferencesFile) {
-    throw new Error(
-        `Can't find a preference file pointer in profile ${me.doc()}`
-    );
-    }
-
-    // //// Load preference file
-    try {
-    await solidLogicSingleton.store.fetcher.load(preferencesFile as NamedNode, {
-        withCredentials: true,
-    });
-    } catch (err) {
-    // Really important to look at why
-    const status = err.status;
-    debug.log(`HTTP status ${status} for preference file ${preferencesFile}`);
-    if (status === 401) {
-        throw new UnauthorizedError();
-    }
-    if (status === 403) {
-        if (differentOrigin()) {
-        throw new CrossOriginForbiddenError();
-        }
-        throw new SameOriginForbiddenError();
-    }
-    if (status === 404) {
-        throw new NotFoundError(preferencesFile.value);
-    }
-    throw new FetchError(err.status, err.message);
-    }
-    return preferencesFile as NamedNode;
-}
 
 export async function ensureLoadedPreferences (context: AuthenticationContext) {
     if (!context.me) throw new Error('@@ ensureLoadedPreferences: no user specified')
@@ -80,39 +19,124 @@ export async function ensureLoadedPreferences (context: AuthenticationContext) {
     return context
 }
 
-export async function loadPreferencesNEW (user: NamedNode): Promise <NamedNode | undefined> {
-    await loadProfileNEW(user)
+/**
+ * loads the preference without throwing errors - if it can create it it does so.
+ * remark: it still throws error if it cannot load profile.
+ * @param user 
+ * @returns undefined if preferenceFile cannot be returned or NamedNode if it can find it or create it
+ */
+export async function loadPreferences (user: NamedNode): Promise <NamedNode | undefined> {
+    await loadProfile(user)
 
-    const possiblePreferencesFile = solidLogicSingleton.profile.suggestPreferencesFile(user)
+    const possiblePreferencesFile = suggestPreferencesFile(user)
 
-    const preferencesFile = await followOrCreateLink(user, ns.space('preferencesFile') as NamedNode, possiblePreferencesFile, user.doc())
-
-    if (!preferencesFile) {
-      const message = `User ${user} has no pointer in profile to preferences file.`
-      console.warn(message)
-      return undefined
+    let preferencesFile
+    try {
+        preferencesFile = await followOrCreateLink(user, ns.space('preferencesFile') as NamedNode, possiblePreferencesFile, user.doc())
+    } catch (err) {
+        const message = `User ${user} has no pointer in profile to preferences file.`
+        debug.warn(message)
+        return undefined 
     }
+
     let response
     try {
         response = await solidLogicSingleton.store.fetcher.load(preferencesFile as NamedNode)
     } catch (err) { // Maybe a permission problem or origin problem
-      console.warn(err.response.status)
-      console.warn(`Unable to load preference of user ${user}: ${err}`)
-      return undefined
+        debug.warn(err.response.status)
+        debug.warn(`Unable to load preference of user ${user}: ${err}`)
+        return undefined
     }
     return preferencesFile as NamedNode
 }
-  
-export async function loadProfileNEW (user: NamedNode):Promise <NamedNode> {
+
+/**
+ * loads the preference without returning different errors if it cannot create or load it.
+ * remark: it also throws error if it cannot load profile.
+ * @param user 
+ * @returns undefined if preferenceFile cannot be an Error or NamedNode if it can find it or create it
+ */
+export async function loadPreferencesThrowErrors (user: NamedNode): Promise <NamedNode> {
+    await loadProfile(user)
+
+    const possiblePreferencesFile = suggestPreferencesFile(user)
+    let preferencesFile
+    try {
+        preferencesFile = await followOrCreateLink(user, ns.space('preferencesFile') as NamedNode, possiblePreferencesFile, user.doc())
+    } catch (err) {
+        const message = `User ${user} has no pointer in profile to preferences file.`
+        debug.warn(message)
+        // we are listing the possible errors
+        if (err instanceof NotEditableError) { throw err }
+        if (err instanceof WebOperationError) { throw err }
+        if (err instanceof UnauthorizedError) { throw err }
+        if (err instanceof CrossOriginForbiddenError) { throw err }
+        if (err instanceof SameOriginForbiddenError) { throw err }
+        if (err instanceof FetchError) { throw err }
+        throw err
+    }
+
+    let response
+    try {
+        response = await solidLogicSingleton.store.fetcher.load(preferencesFile as NamedNode)
+    } catch (err) { // Maybe a permission problem or origin problem
+        const msg = `Unable to load preference of user ${user}: ${err}`
+        debug.warn(msg)
+        if (err.response.status === 401) {
+            throw new UnauthorizedError();
+        }
+        if (err.response.status === 403) {
+            if (differentOrigin(preferencesFile)) {
+            throw new CrossOriginForbiddenError();
+            }
+            throw new SameOriginForbiddenError();
+        }
+        if (err.response.status === 404) {
+            throw new NotFoundError();
+        }
+        throw new Error(msg)
+    }
+    return preferencesFile as NamedNode
+}
+
+export async function loadProfile (user: NamedNode):Promise <NamedNode> {
     if (!user) {
-      throw new Error(`loadProfile: no user given.`)
+        throw new Error(`loadProfile: no user given.`)
     }
     try {
-      await solidLogicSingleton.store.fetcher.load(user.doc())
+        await solidLogicSingleton.store.fetcher.load(user.doc())
     } catch (err) {
-      throw new Error(`Unable to load profile of user ${user}: ${err}`)
+        throw new Error(`Unable to load profile of user ${user}: ${err}`)
     }
     return user.doc()
-  }
+}
 
+export async function loadMe(): Promise<NamedNode> {
+    const me = solidLogicSingleton.authn.currentUser();
+    if (me === null) {
+        throw new Error("Current user not found! Not logged in?");
+    }
+    await store.fetcher.load(me.doc());
+    return me;
+}
 
+export function getPodRoot(user: NamedNode): NamedNode {
+    const podRoot = findStorage(user);
+    if (!podRoot) {
+        throw new Error("User pod root not found!");
+    }
+    return podRoot as NamedNode;
+}
+
+export async function getMainInbox(user: NamedNode): Promise<NamedNode> {
+    await store.fetcher.load(user);
+    const mainInbox = store.any(user, ns.ldp("inbox"), undefined, user.doc());
+    if (!mainInbox) {
+        throw new Error("User main inbox not found!");
+    }
+    return mainInbox as NamedNode;
+}
+
+export function findStorage(me: NamedNode) {
+    return store.any(me, ns.space("storage"), undefined, me.doc());
+}
