+++ src/typeIndex/typeIndexLogic.ts
@@ -1,15 +1,15 @@
+import * as $rdf from 'rdflib'
 import * as rdf from "rdflib"
-import { NamedNode, st } from 'rdflib'
+import { NamedNode, st, sym } from 'rdflib'
+import solidNamespace from 'solid-namespace'
 import { solidLogicSingleton } from "../logic/solidLogicSingleton"
+import { ensureLoadedPreferences, loadPreferences, loadProfile } from "../profile/profileLogic"
 import { AuthenticationContext, ScopedApp, TypeIndexScope } from '../types'
 import * as debug from "../util/debug"
-import solidNamespace from 'solid-namespace'
-import * as $rdf from 'rdflib'
-import { ensureLoadedPreferences, loadPreferences, loadPreferencesNEW, loadProfileNEW } from "../profile/profileLogic"
-import { suggestPrivateTypeIndex, suggestPublicTypeIndex } from "../discovery/discoveryLogic"
-import { followOrCreateLink } from "../util/utilityLogic"
-import { newThing } from "../util/utils"
+import { createEmptyRdfDoc, followOrCreateLink } from "../util/utilityLogic"
+import { newThing, uniqueNodes } from "../util/utils"
 
+const store = solidLogicSingleton.store
 const ns = solidNamespace($rdf)
 /**
  * Resolves with the same context, outputting
@@ -193,14 +193,6 @@ async function makeIndexIfNecessary (context, isPublic, store, ns) {
     }
 } // makeIndexIfNecessary
 
-async function createEmptyRdfDoc(doc: NamedNode, comment: string) {
-    await solidLogicSingleton.store.fetcher.webOperation("PUT", doc.uri, {
-    data: `# ${new Date()} ${comment}
-`,
-    contentType: "text/turtle",
-    });
-}
-
 async function loadIndexes(
     me: NamedNode | string,
     publicProfile: NamedNode | string | null,
@@ -257,7 +249,7 @@ function getTypeIndex(
     preferencesFile: NamedNode | string,
     isPublic: boolean
 ): NamedNode[] {
-    // console.log('getTypeIndex', this.store.each(me, undefined, undefined, preferencesFile), isPublic, preferencesFile)
+    // console.log('getTypeIndex', store.each(me, undefined, undefined, preferencesFile), isPublic, preferencesFile)
     return solidLogicSingleton.store.each(
     me as NamedNode,
     isPublic ? ns.solid("publicTypeIndex") : ns.solid("privateTypeIndex"),
@@ -274,63 +266,75 @@ function getRegistrations(instance, theClass) {
     });
 }
 
-//------- NEW
 async function loadTypeIndexesFor (user: NamedNode): Promise<Array<TypeIndexScope>> {
     if (!user) throw new Error(`loadTypeIndexesFor: No user given`)
-    const profile = await loadProfileNEW(user)
+    const profile = await loadProfile(user)
 
     const suggestion = suggestPublicTypeIndex(user)
-
-    const publicTypeIndex = await followOrCreateLink(user, ns.solid('publicTypeIndex') as NamedNode, suggestion, profile)
-
+    let publicTypeIndex
+    try {
+        publicTypeIndex = await followOrCreateLink(user, ns.solid('publicTypeIndex') as NamedNode, suggestion, profile)
+    } catch (err) {
+        const message = `User ${user} has no pointer in profile to publicTypeIndex file.`
+        debug.warn(message)
+    }
     const publicScopes = publicTypeIndex ? [ { label: 'public', index: publicTypeIndex as NamedNode, agent: user } ] : []
 
     let preferencesFile
     try {
-      preferencesFile = await loadPreferencesNEW(user)
+        preferencesFile = await loadPreferences(user)
     } catch (err) {
-      preferencesFile = null
+        preferencesFile = null
     }
 
     let privateScopes
     if (preferencesFile) { // watch out - can be in either as spec was not clear.  Legacy is profile.
       // If there is a legacy one linked from the profile, use that.
       // Otherwiae use or make one linked from Preferences
-      const suggestedPrivateTypeIndex = suggestPrivateTypeIndex(preferencesFile)
-
-      const privateTypeIndex = solidLogicSingleton.store.any(user, ns.solid('privateTypeIndex'), undefined, profile) ||
-
-      await followOrCreateLink(user, ns.solid('privateTypeIndex') as NamedNode, suggestedPrivateTypeIndex, preferencesFile);
-
-      privateScopes = privateTypeIndex ? [ { label: 'private', index: privateTypeIndex as NamedNode, agent: user } ] : []
+        const suggestedPrivateTypeIndex = suggestPrivateTypeIndex(preferencesFile)
+        let privateTypeIndex
+        try {
+            privateTypeIndex = solidLogicSingleton.store.any(user, ns.solid('privateTypeIndex'), undefined, profile) ||
+                await followOrCreateLink(user, ns.solid('privateTypeIndex') as NamedNode, suggestedPrivateTypeIndex, preferencesFile);
+        } catch (err) {
+            const message = `User ${user} has no pointer in preference file to privateTypeIndex file.`
+            debug.warn(message)
+        }
+        privateScopes = privateTypeIndex ? [ { label: 'private', index: privateTypeIndex as NamedNode, agent: user } ] : []
     } else {
-      privateScopes = []
+        privateScopes = []
     }
     const scopes = publicScopes.concat(privateScopes)
     if (scopes.length === 0) return scopes
     const files = scopes.map(scope => scope.index)
     try {
-      await solidLogicSingleton.store.fetcher.load(files)
+        await solidLogicSingleton.store.fetcher.load(files)
     } catch (err) {
-      console.warn('Problems loading type index: ', err)
+        debug.warn('Problems loading type index: ', err)
     }
     return scopes
 }
 
 async function loadCommunityTypeIndexes (user: NamedNode): Promise<TypeIndexScope[][]> {
-    const preferencesFile = await loadPreferencesNEW(user)
+    let preferencesFile
+    try {
+        preferencesFile = await loadPreferences(user)
+    } catch (err) {
+        const message = `User ${user} has no pointer in profile to preferences file.`
+        debug.warn(message)
+    }
     if (preferencesFile) { // For now, pick up communities as simple links from the preferences file.
-      const communities = solidLogicSingleton.store.each(user, ns.solid('community'), undefined, preferencesFile as NamedNode).concat(
+        const communities = solidLogicSingleton.store.each(user, ns.solid('community'), undefined, preferencesFile as NamedNode).concat(
         solidLogicSingleton.store.each(user, ns.solid('community'), undefined, user.doc() as NamedNode)
-      )
-      let result = []
-      for (const org of communities) {
-        result = result.concat(await loadTypeIndexesFor(org as NamedNode) as any)
-      }
-      return result
+        )
+        let result = []
+        for (const org of communities) {
+            result = result.concat(await loadTypeIndexesFor(org as NamedNode) as any)
+        }
+        return result
     }
     return [] // No communities
-  }
+}
 
 async function loadAllTypeIndexes (user: NamedNode) {
     return (await loadTypeIndexesFor(user)).concat((await loadCommunityTypeIndexes(user)).flat())
@@ -340,7 +344,7 @@ async function getScopedAppInstances (klass: NamedNode, user: NamedNode):Promise
     const scopes = await loadAllTypeIndexes(user)
     let scopedApps = []
     for (const scope of scopes) {
-        const scopedApps0 = await solidLogicSingleton.typeIndex.getScopedAppsFromIndex(scope, klass) as any
+        const scopedApps0 = await getScopedAppsFromIndex(scope, klass) as any
         scopedApps = scopedApps.concat(scopedApps0)
     }
     return scopedApps
@@ -355,16 +359,95 @@ async function getAppInstances (klass: NamedNode): Promise<NamedNode[]> {
     const scopedAppInstances = await getScopedAppInstances(klass, user)
     return scopedAppInstances.map(scoped => scoped.instance)
 }
-  
+
+function suggestPublicTypeIndex (me: NamedNode) {
+    return sym(me.doc().dir()?.uri + 'publicTypeIndex.ttl')
+}
+// Note this one is based off the pref file not the profile
+
+function suggestPrivateTypeIndex (preferencesFile: NamedNode) {
+    return sym(preferencesFile.doc().dir()?.uri + 'privateTypeIndex.ttl')
+}
+
+/*
+* Register a new app in a type index
+* used in chat in bookmark.js (solid-ui)
+* Returns the registration object if successful else null
+*/
+async function registerInstanceInTypeIndex (
+instance: NamedNode,
+index: NamedNode,
+theClass: NamedNode,
+// agent: NamedNode
+): Promise<NamedNode | null> {
+const registration = newThing(index)
+    const ins = [
+        // See https://github.com/solid/solid/blob/main/proposals/data-discovery.md
+        st(registration, ns.rdf('type'), ns.solid('TypeRegistration'), index),
+        st(registration, ns.solid('forClass'), theClass, index),
+        st(registration, ns.solid('instance'), instance, index)
+    ]
+    try {
+        await store.updater.update([], ins)
+    } catch (err) {
+        const msg = `Unable to register ${instance} in index ${index}: ${err}`
+        console.warn(msg)
+        return null
+    }
+    return registration
+}
+
+async function deleteTypeIndexRegistration (item) {
+    const reg = store.the(null, ns.solid('instance'), item.instance, item.scope.index) as NamedNode
+    if (!reg) throw new Error(`deleteTypeIndexRegistration: No registration found for ${item.instance}`)
+    const statements = store.statementsMatching(reg, null, null, item.scope.index)
+    await store.updater.update(statements, [])
+}
+
+async function getScopedAppsFromIndex (scope, theClass: NamedNode | null) {
+    const index = scope.index
+    const registrations = store.statementsMatching(null, ns.solid('instance'), null, index)
+        .concat(store.statementsMatching(null, ns.solid('instanceContainer'), null, index))
+        .map(st => st.subject)
+    const relevant = theClass ? registrations.filter(reg => store.any(reg, ns.solid('forClass'), null, index)?.sameTerm(theClass))
+        : registrations
+    const directInstances = relevant.map(reg => store.each(reg, ns.solid('instance'), null, index).map(one => sym(one.value))).flat()
+    let instances = uniqueNodes(directInstances)
+
+    const instanceContainers = relevant.map(
+        reg => store.each(reg, ns.solid('instanceContainer'), null, index).map(one => sym(one.value))).flat()
+
+    //  instanceContainers may be deprocatable if no one has used them
+    const containers = uniqueNodes(instanceContainers)
+    if (containers.length > 0) { console.log('@@ getScopedAppsFromIndex containers ', containers)}
+    for (let i = 0; i < containers.length; i++) {
+        const cont = containers[i]
+        await store.fetcher.load(cont)
+        const contents = store.each(cont, ns.ldp('contains'), null, cont).map(one => sym(one.value))
+        instances = instances.concat(contents)
+    }
+    return instances.map(instance => { return {instance, scope}})
+}
+
 export {
     ensureTypeIndexes,
     loadTypeIndexes,
-    getRegistrations,
     registerInTypeIndex,
     loadIndex,
+    ensureOneTypeIndex,
+    putIndex,
+    makeIndexIfNecessary,
+    loadIndexes,
+    getTypeIndex,
+    getRegistrations,
     loadTypeIndexesFor,
     loadCommunityTypeIndexes,
     loadAllTypeIndexes,
     getScopedAppInstances,
-    getAppInstances
+    getAppInstances,
+    suggestPublicTypeIndex,
+    suggestPrivateTypeIndex,
+    registerInstanceInTypeIndex,
+    deleteTypeIndexRegistration,
+    getScopedAppsFromIndex
 }
